[fe] refactor out models/
