+++ web-client/src/models/resource.ts
@@ -1,0 +1,6 @@
+export interface Resource {
+    id: string;
+    name: string;
+    url: string;
+    previewUrl?: string;
+}
