+++ web-client/src/models/tag.ts
@@ -1,0 +1,7 @@
+import { Resource } from './resource';
+
+export interface Tag {
+    id: string;
+    name: string;
+    resource?: string | Resource;
+}
